webapp/requirements.txt
Flask
paho-mqtt
python-dotenv
pandas
plotly
pytz